# wasi: fixes fd_readdir bench