The benchmark for `fd_readdir` has been broken for a while. We need to fix this before releasing to ensure we know the performance after various recent work.

I ran this benchmark now vs 34324031cb6d1f4b96ff706c72acd054fc70d6fd, which was before we started changing the readdir impl.

A later PR will correct the performance by being more like the former <1.2.0 impl.

```
goos: darwin
goarch: arm64
pkg: github.com/tetratelabs/wazero/imports/wasi_snapshot_preview1
                                      │ before-refactor.txt │              now.txt               │
                                      │       sec/op        │   sec/op     vs base               │
_fdReaddir/embed.FS-12                          2.255µ ± 7%   2.879µ ± 6%  +27.70% (p=0.002 n=6)
_fdReaddir/embed.FS_-_two_calls-12              2.897µ ± 1%   3.606µ ± 3%  +24.50% (p=0.002 n=6)
_fdReaddir/os.DirFS-12                          42.74µ ± 1%   43.62µ ± 1%   +2.07% (p=0.002 n=6)
_fdReaddir/os.DirFS_-_two_calls-12              43.24µ ± 0%   44.15µ ± 2%   +2.12% (p=0.002 n=6)
_fdReaddir/sysfs.DirFS-12                       39.71µ ± 0%   41.33µ ± 0%   +4.08% (p=0.002 n=6)
_fdReaddir/sysfs.DirFS_-_two_calls-12           40.67µ ± 1%   41.57µ ± 2%   +2.21% (p=0.002 n=6)
geomean                                         16.40µ        18.03µ        +9.91%

                                      │ before-refactor.txt │               now.txt               │
                                      │        B/op         │     B/op      vs base               │
_fdReaddir/embed.FS-12                         1.250Ki ± 0%   2.344Ki ± 0%  +87.50% (p=0.002 n=6)
_fdReaddir/embed.FS_-_two_calls-12             1.359Ki ± 0%   2.609Ki ± 0%  +91.95% (p=0.002 n=6)
_fdReaddir/os.DirFS-12                         6.141Ki ± 0%   8.391Ki ± 0%  +36.64% (p=0.002 n=6)
_fdReaddir/os.DirFS_-_two_calls-12             7.445Ki ± 0%   8.672Ki ± 0%  +16.47% (p=0.002 n=6)
_fdReaddir/sysfs.DirFS-12                      5.906Ki ± 0%   8.211Ki ± 0%  +39.02% (p=0.002 n=6)
_fdReaddir/sysfs.DirFS_-_two_calls-12          7.211Ki ± 0%   8.492Ki ± 0%  +17.77% (p=0.002 n=6)
geomean                                        3.860Ki        5.606Ki       +45.22%

                                      │ before-refactor.txt │              now.txt              │
                                      │      allocs/op      │ allocs/op   vs base               │
_fdReaddir/embed.FS-12                           13.00 ± 0%   19.00 ± 0%  +46.15% (p=0.002 n=6)
_fdReaddir/embed.FS_-_two_calls-12               18.00 ± 0%   24.00 ± 0%  +33.33% (p=0.002 n=6)
_fdReaddir/os.DirFS-12                           56.00 ± 0%   66.00 ± 0%  +17.86% (p=0.002 n=6)
_fdReaddir/os.DirFS_-_two_calls-12               64.00 ± 0%   71.00 ± 0%  +10.94% (p=0.002 n=6)
_fdReaddir/sysfs.DirFS-12                        51.00 ± 0%   62.00 ± 0%  +21.57% (p=0.002 n=6)
_fdReaddir/sysfs.DirFS_-_two_calls-12            59.00 ± 0%   67.00 ± 0%  +13.56% (p=0.002 n=6)
geomean                                          36.90        45.50       +23.32%
```