@codefromthecrypt 2023-07-04
It takes a while to run this, so I'll comment back on perf since before we started refactoring.